feat: add seed to remove randomness
